projects/pjsekai-bot/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "pjsekai-bot"
version = "0.0.1.dev0"
description = "pjsekai bot."
license = "MIT"
authors = [
    "TheerapakG <theerapakg@gmail.com>",
]

[tool.poetry.dependencies]
python = "^3.8"

"discord.py" = "^2.2.2"
GitPython = "^3.1.11"
jycm = "^1.5.0"
python-dateutil = "^2.8.2"
python-dotenv = "^1.0.0"
UnityPy = "^1.9.28"
async-pjsekai = { path="../../libs/async-pjsekai", develop=true }
